+++ examples/custom-types/expected/Operations/MyCustomEnumQuery/MyCustomEnumQuery.php
@@ -13,7 +13,9 @@ class MyCustomEnumQuery extends \Spawnia\Sailor\ObjectLike
     /**
      * @param \Spawnia\Sailor\CustomTypes\Types\CustomEnum|null $withCustomEnum
      */
-    public static function make($withCustomEnum = 'Special default value that allows Sailor to differentiate between explicitly passing null and not passing a value at all.'): self
+    public static function make(
+        $withCustomEnum = 'Special default value that allows Sailor to differentiate between explicitly passing null and not passing a value at all.',
+    ): self
     {
         $instance = new self;