+++ examples/custom-types/expected/Operations/MyNativeEnumQuery/MyNativeEnumQueryErrorFreeResult.php
@@ -1,0 +1,20 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Spawnia\Sailor\CustomTypes\Operations\MyNativeEnumQuery;
+
+class MyNativeEnumQueryErrorFreeResult extends \Spawnia\Sailor\ErrorFreeResult
+{
+    public MyNativeEnumQuery $data;
+
+    public static function endpoint(): string
+    {
+        return 'custom-types';
+    }
+
+    public static function config(): string
+    {
+        return \Safe\realpath(__DIR__ . '/../../../sailor.php');
+    }
+}