+++ examples/custom-types/sailor.php
@@ -10,6 +10,7 @@ use Spawnia\Sailor\EndpointConfig;
 use Spawnia\Sailor\Response;
 use Spawnia\Sailor\Testing\MockClient;
 use Spawnia\Sailor\Type\BenSampoEnumTypeConfig;
+use Spawnia\Sailor\Type\NativeEnumTypeConfig;
 
 return [
     'custom-types' => new class() extends EndpointConfig {
@@ -71,6 +72,7 @@ return [
                     'CustomDate' => new CustomDateTypeConfig($this, $schema->getType('CustomDate')),
                     'CustomInput' => new CustomObjectTypeConfig($this, $schema->getType('CustomInput')),
                     'CustomOutput' => new CustomObjectTypeConfig($this, $schema->getType('CustomOutput')),
+                    'NativeEnum' => new NativeEnumTypeConfig($this, $schema->getType('NativeEnum')),
                 ]
             );
         }