+++ examples/polymorphic/expected/Operations/PolymorphicCommonSubChildren/Sub/Nodes/Post.php
@@ -20,8 +20,9 @@ class Post extends \Spawnia\Sailor\ObjectLike
     public static function make(
         $id,
         $node = 'Special default value that allows Sailor to differentiate between explicitly passing null and not passing a value at all.',
-        $title = 'Special default value that allows Sailor to differentiate between explicitly passing null and not passing a value at all.'
-    ): self {
+        $title = 'Special default value that allows Sailor to differentiate between explicitly passing null and not passing a value at all.',
+    ): self
+    {
         $instance = new self;
 
         if ($id !== self::UNDEFINED) {