+++ examples/simple/expected/Operations/MyObjectNestedQuery/MyObjectNestedQuery.php
@@ -13,7 +13,9 @@ class MyObjectNestedQuery extends \Spawnia\Sailor\ObjectLike
     /**
      * @param \Spawnia\Sailor\Simple\Operations\MyObjectNestedQuery\SingleObject\SomeObject|null $singleObject
      */
-    public static function make($singleObject = 'Special default value that allows Sailor to differentiate between explicitly passing null and not passing a value at all.'): self
+    public static function make(
+        $singleObject = 'Special default value that allows Sailor to differentiate between explicitly passing null and not passing a value at all.',
+    ): self
     {
         $instance = new self;