+++ examples/simple/expected/Operations/MyObjectNestedQuery/SingleObject/SomeObject.php
@@ -13,7 +13,9 @@ class SomeObject extends \Spawnia\Sailor\ObjectLike
     /**
      * @param \Spawnia\Sailor\Simple\Operations\MyObjectNestedQuery\SingleObject\Nested\SomeObject|null $nested
      */
-    public static function make($nested = 'Special default value that allows Sailor to differentiate between explicitly passing null and not passing a value at all.'): self
+    public static function make(
+        $nested = 'Special default value that allows Sailor to differentiate between explicitly passing null and not passing a value at all.',
+    ): self
     {
         $instance = new self;